Changed some params to double in Calculate class

+++ ATMHandin3/ATMHandin3/Classes/Calculate.cs
@@ -5,14 +5,14 @@ namespace ATMHandin3.Classes
 {
     public static class Calculate
     {
-        public static double CalculateVelocity(int xcoor1, int xcoor2, int ycoor1, int ycoor2, DateTime timestamp1, DateTime timestamp2)
+        public static double CalculateVelocity(double xcoor1, double xcoor2, double ycoor1, double ycoor2, DateTime timestamp1, DateTime timestamp2)
         {
             double xdiff = xcoor2 - xcoor1;
             double ydiff = ycoor2 - ycoor1;
 
             double distance = Math.Sqrt(Math.Pow(xdiff, 2) + Math.Pow(ydiff, 2));
 
-            double velocity = (int)(distance / (timestamp2 - timestamp1).TotalSeconds);
+            double velocity = (distance / (timestamp2 - timestamp1).TotalSeconds);
 
             return velocity;
         }
@@ -44,9 +44,9 @@ namespace ATMHandin3.Classes
             return longtitude;
         }
 
-        public static int CalculateAltitudeDiff(int alti1, int alti2)
+        public static double CalculateAltitudeDiff(double alti1, double alti2)
         {
-            int diffAlti;
+            double diffAlti;
 
             if (alti1 > alti2)
             {
